+++ pepper/rc/changelog.md
@@ -26,6 +26,7 @@
 - changed `plugin-remedybg` to make use of the new IPC driver api which enables better integration
 - changed `help` command to autocomplete on help page names and always open help pages
 - fix web version not working because it would try to create a log file
+- changed cursor list format from "path:1,3-1,4;2,3-2,4" to "path:1:3-1:4,2:3-2:4"
 
 # 0.29.0
 - added diagnostic logging to `spawn` command